--- src/stores/search-flag-store.ts
@@ -1,11 +1,0 @@
-import { create } from "zustand";
-
-type SearchFlagState = {
-  searchFlag: boolean;
-  setSearchFlag: (flag: boolean) => void;
-};
-
-export const useSearchFlagStore = create<SearchFlagState>()((set) => ({
-  searchFlag: false,
-  setSearchFlag: (flag) => set(() => ({ searchFlag: flag })),
-}));
